Rename folders and files.
