+++ client/src/App.js
@@ -1,7 +1,7 @@
 import React, {useEffect} from 'react'
 import 'antd/dist/antd.less'
 import { Layout, Menu } from 'antd';
-import SiteHeader from './components/header/SiteHeader'
+import SiteHeader from './components/Header'
 import { getProducts } from './functions/products/product'
 import MainRoutes from './routes/MainRoutes'
 const { Content, Footer } = Layout;
