+++ client/src/pages/Cart/index.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import './Cart.less'
+import './styles.less'
 
 function Cart (props) {
   return (
