+++ client/src/serviceWorker.js
@@ -1,5 +1,5 @@
-// This optional code is used to register a service worker.
-// register() is not called by default.
+// This optional code is used to Register a service worker.
+// Register() is not called by default.
 
 // This lets the app load faster on subsequent visits in production, and gives
 // it offline capabilities. However, it also means that developers (and users)
@@ -47,7 +47,7 @@ export function register(config) {
           );
         });
       } else {
-        // Is not localhost. Just register service worker
+        // Is not localhost. Just Register service worker
         registerValidSW(swUrl, config);
       }
     });
