+++ routes/paymentMethods.js
@@ -16,7 +16,7 @@ const {
 // @access  Private
 router.post(
   "/",
-  passport.authenticate("jwt-admin", { session: false }),
+  passport.authenticate("jwt-Admin", { session: false }),
   addPaymentMethod
 );
 
@@ -25,7 +25,7 @@ router.post(
 // @access  Private
 router.put(
   "/:customId",
-  passport.authenticate("jwt-admin", { session: false }),
+  passport.authenticate("jwt-Admin", { session: false }),
   updatePaymentMethod
 );
 
@@ -34,7 +34,7 @@ router.put(
 // @access  Private
 router.delete(
   "/:customId",
-  passport.authenticate("jwt-admin", { session: false }),
+  passport.authenticate("jwt-Admin", { session: false }),
   deletePaymentMethod
 );
